Refactor session management and improve usability

+++ src/main/java/ro/unibuc/hello/controller/ItemsController.java
@@ -1,8 +1,8 @@
 package ro.unibuc.hello.controller;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,12 +12,13 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import ro.unibuc.hello.dto.Item;
-import ro.unibuc.hello.dto.ItemPost;
-import ro.unibuc.hello.exception.EntityNotFoundException;
+import jakarta.servlet.http.HttpServletRequest;
+import ro.unibuc.hello.auth.AuthUtil;
+import ro.unibuc.hello.auth.PublicEndpoint;
+import ro.unibuc.hello.dto.ItemPostRequest;
+import ro.unibuc.hello.dto.ItemWithOwner;
+import ro.unibuc.hello.permissions.ItemPermissionChecker;
 import ro.unibuc.hello.service.ItemsService;
-
-import java.util.List;
 
 @Controller
 public class ItemsController {
@@ -25,43 +26,43 @@ public class ItemsController {
     @Autowired
     private ItemsService itemsService;
 
+    @Autowired
+    private ItemPermissionChecker permissionChecker;
+
+    @PublicEndpoint
     @GetMapping("/items")
     @ResponseBody
-    public List<Item> getAllItems() {
+    public List<ItemWithOwner> getAllItems() {
         return itemsService.getAllItems();
+    }
+
+    @PublicEndpoint
+    @GetMapping("/items/{id}")
+    @ResponseBody
+    public ItemWithOwner getItemById(@PathVariable String id) {
+        return itemsService.getItemById(id);
     }
 
     @PostMapping("/items")
     @ResponseBody
-    public Item createItem(@RequestBody ItemPost item) {
-        return itemsService.saveItem(item);
-    }
-
-    @GetMapping("/items/{id}")
-    public ResponseEntity<Item> getItemById(@PathVariable String id) {
-          
-        Item item = itemsService.getItemById(id);
-
-        if (item != null) {
-          return new ResponseEntity<>(item, HttpStatus.OK);
-        } else {
-          return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-
+    public ItemWithOwner createItem(HttpServletRequest request, @RequestBody ItemPostRequest item) {
+        String userId = AuthUtil.getAuthenticatedUserId(request);
+        return itemsService.saveItem(userId, item);
     }
 
     @PutMapping("/items/{id}")
     @ResponseBody
-    public Item updateItem(@PathVariable String id, @RequestBody Item item) throws EntityNotFoundException {
+    public ItemWithOwner updateItem(HttpServletRequest request, @PathVariable String id, @RequestBody ItemPostRequest item) {
+        String userId = AuthUtil.getAuthenticatedUserId(request);
+        permissionChecker.checkOwnership(userId, id);
         return itemsService.updateItem(id, item);
     }
 
-
     @DeleteMapping("/items/{id}")
     @ResponseBody
-    public void deleteItem(@PathVariable String id) throws EntityNotFoundException {
-          itemsService.deleteItem(id);
+    public void deleteItem(HttpServletRequest request, @PathVariable String id) {
+        String userId = AuthUtil.getAuthenticatedUserId(request);
+        permissionChecker.checkOwnership(userId, id);
+        itemsService.deleteItem(id);
     }
-
 }
-
